Allow to configure unlimited number of tools
Sometimes we need to use some tools several times with different configuration.
Sometimes we don't need some tools at all.
And finally sometimes we need to compare tools reports in different ways.
For example we can run benchmark with same code but with another argument and wanted to analyze progression.
But with another tool we want to compare detailed results of same code run with same argument.
This commit allow to configure tools and display of comparison table.
Also this is a first attempt to stabilize gem API.